Move setup.py up a folder because it's the only file in a folder
