+++ packages/grid/apps/domain/src/main/core/services/setup_service.py
@@ -28,7 +28,7 @@ from syft.grid.messages.setup_messages import GetSetUpResponse
 # from ...core.infrastructure import AWS_Serverless
 # from ...core.infrastructure import Config
 # from ...core.infrastructure import Provider
-from ..database.setup.setup import SetupConfig
+from ..database.setup import SetupConfig
 from ..database.utils import model_to_json
 from ..exceptions import AuthorizationError
 from ..exceptions import InvalidParameterValueError
